System resolves Rethink ingestion altogether #326

+++ backend/dhri/install_parser.py
@@ -24,11 +24,13 @@ class InstallCache():
         self.path = CACHE_DIRS['ROOT'] / (self.loader.repo_name + ".json")
         self.expired = _is_expired(self.path, age_checker=TEST_AGES["INSTALL"], force_download=force_download) == True
 
+        new_content = False
         if not self.path.exists():
             self.log.warning(f'{self.path} does not exist so downloading install cache...')
             self._setup_raw_content()
+            new_content = True
 
-        if force_download == True or self.expired == True:
+        if new_content == False and (force_download == True or self.expired == True):
             self.log.warning(f'Force download is set to True or cache file has expired so downloading install cache...')
             self._setup_raw_content()
